requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
djangorestframework==3.15.2
djongo==1.3.7
pycparser==2.22
pymongo==3.11.4
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.1
